+++ precompiles/src/math/mod.rs
@@ -1,0 +1,11 @@
+//! Arithmetic deferred precompiles.
+
+pub mod curve;
+pub mod ed25519_base;
+pub mod ed25519_scalar;
+pub mod k1_base;
+pub mod k1_scalar;
+pub mod r1_base;
+pub mod r1_scalar;
+pub mod u256;
+pub mod uint;